AuthorizationRequest: can add contacts
Use a combinaison of class method and `ActiveRecord::Store` in order to
manage contacts.

`ActiveRecord::Store` is a thin wrapper around jsonb to deal with
key/value store as others database attributes. Thanks to this approach
we can use validators, and it will be easy to retrieves authorizations
requests linked to emails.

And please.. look at this beautiful interface on
`AuthorizationRequest::HubEECertDC`, it's tasty!