Rewrite all logic for models
Introduce an AuthorizationDefinition, which bundle all definitions of
specific authorization. They were previously on forms, but it's not
accurate if you want to have multiple forms: some attribute like unique,
name, description are specific to the authorization, not the form.

In the previous version it was not an issue because 1 form = 1
authorization, but if you want to introduce N forms for 1
authorization, you have to make the switch.

This commit prepare this to have a page with all possible authorizations
("habilitations"), when you want to start one, you have 2 cases:

1. There's only one form -> redirect to it
2. There is at least 2 forms -> ask which one (and maybe more ?)

Currently 2. is not implemented.

Simplify authorization_request_form helper too