Harden tests
spec/features/profile_spec.rb failed with seed 9503. After inspect it
seems that OmniAuth was not in test mode.

There was a specific config for request specs on OmniAuth, which seems
to be useless: `OmniAuth.config.test_mode` is already true thanks to
initializers, and remove the `mon_compte_pro` mock is not relevant on
all tests.

This commit simplifies and make it green