Enhance scopes management for changelogs
There's multiple cases where scopes messed up with changelogs:

* When scopes are included by default (from definition)
* When scopes are included thanks to form
* Combinaison of both of them

When user creates the authorization request or updates scopes through
wizard or single page form, scopes used to be changed in a lot of
unconsistent ways.

Because of these inconsistences, changelogs were really messy and false.

This commit tries to harmonize of we manipulate scopes in order to keep
it consistent across create/update, to displays an accurate changelog
based on actual change of user and not of the random order applied by
view/server or database.

This is achieved thanks to:

1. Attr reader/writer harden and normalize: we cast as an array and sort
   everytime we can, to keep it consistent
2. Use attr reader everywhere instead of `data` key, in order to
   properly use 1.
3. AuthorizationRequestForm#data adds included scopes from definition
   for `scopes` key, to keep it consistent too