Starting a form no longers persist it in database
This commit change how we manage authorization request lifecycle.
Previously when a user choose a form (from decision tree, index or
directly through a link), click on start used to persisted the record.

Now we arrive on a proper "new" action/view (which is start, new is keeped
for direct link), and the record is only saved on the first save/next.

Some changelogs:

* Do not show badge number (because there is no ID yet)
* Create displays an alert on success
* TODO Display an info alert to inform the user that it is a draft