Harden CreateAuthorizationRequestChangelog by using attr reader instead of raw data key
Closes https://errors.data.gouv.fr/organizations/sentry/issues/140124/

Before this version, we used to compare, for a key `k`,
`authorization_request.send(k)` and `authorization_request.data[k]`.

Because of some override, for example `scopes`, we used to create diff
which are not relevant, and because of the difference on `scopes` between the
reader (which renders an array) and the `data` (which is a string), our
changelog builder raised the issue reported at the beginning of this
commit. We used to store the following within `diff`:

```
{
  "scope": ["['scope1', 'scope2']", ['scope1', 'scope2']]
}
```

This commit introduces a normalisation, and build a proper changelog to
display.