Dynamically set host from mailer params
Because of the introduction of API Particulier with a subdomain we have to
set host dynamically. Thanks to subdomains config model we can fetch the
url with the link between authorization requests and subdomains.

This is a temporary solution, it won't work (well?) with authorization
requests linked to multiple subdomains.

Moreover when the migration from v1 will be over, perhaps we will set
the main domain ? (TBD)

Closes https://linear.app/pole-api/issue/DAT-438