Harden notification settings boolean values
In database when you manually change toggles from settings UI, we store
'0' and '1', not boolean. We have to deal with these values too.

Thanks to this commit users who disabled notification from UI no longer
receive emails.

For information there is users with boolean values on these settings,
because of migration from v1 to v2.