Harden scopes changelog on changed prefilled form
`values[0]` can be `nil`, this method is not evaluated if `name`
is `submit_with_unchanged_prefilled_values`. However if form's scopes
changed since the authorization request creation, we have a diff on
scopes between initial form and authorization request, which leads to
build the scopes change entry.

In order to mitigate 500 errors here, we are taking the form and build a
diff with it : it is not 100% accurate, but better than error 500.

Ref https://linear.app/pole-api/issue/DAT-525/reouverture-dune-demande-dont-certains-scopes-voir-meme-champs-ont#comment-23bb258c