Harden changelog displays for authorization requests with changed prefilled data
When a new authorization request is :

1. submitted for the first time
2. linked to a prefilled form with scopes key
3. without changed data data for this scopes key

We build a changelog with a scopes key with an array as following:
`[nil, ['scope1', 'scope2']]`

When we want to displays it (thanks to
`AuthorizationRequestEventDecorator`), we filter keys that did not
change between the `form.data[key]` and the
`authorization_request.public_send(key)`.

There is one (big) problem here: if we change the `form.data` it will
displays a line for this key, even if the user did not change it.

Moreover, for the `scopes` key, because the key is not filtered, when we
try to build the changelog (through `build_scopes_change` method), we
try to substract an array with nil, which leads to a 500 error.

This commit avoid this situation by removing the key 'scopes' if the
first array entry from the changelog is empty (which is an
initial submit without change on prefilled data).

Ref https://github.com/etalab/data_pass/pull/352 (other solution, but a
bad one)
Ref https://linear.app/pole-api/issue/DAT-525/reouverture-dune-demande-dont-certains-scopes-voir-meme-champs-ont#comment-23bb258c

Closes https://errors.data.gouv.fr/organizations/sentry/issues/143777/