Extensive tests (and some changes) on CreateAuthorizationRequestChangelog
Tests all possible cases on changelog creation.

Moreover we change how we build changelogs:

* The first one stores every keys, with `nil` as first value if there is
  no prefilled data, prefilled data otherwise (new behaviour)
* Next changelogs build diff only with changes keys (same behaviour as
  before)

The `reified_data` within the object build by retrieving all changelog
and first the nearest key to get the value. If it's missing (for
whatever reason), it is excluded and the key is not mark as new value.

This new implementation need 2 things:

1. Change how we display changelogs
2. Handle old data

For 2., we're going to mark all changelogs as "legacy" and displays a
simple sentence instead of trying to rebuild the full changelog. Data
will be migrated step by step, depends of the cases.