Introduce required_personal_email option on contacts
Thanks to this new option on contacts we can trigger the individual-email-checker
stimulus controller on contacts' emails.

This commit introduce a ContactDefinition model, in order to manipulate
the new option. Contact model is still in use in some parts of the code.
We clarify usage between these 2 models as well, because Contact model
was misused.

Closes https://linear.app/pole-api/issue/DAT-552