Removing unicity clause on UUID for MalwareScan model
If 2 people upload the same attachment, we get the same
response from JCOP, with the same UUID.

Because of unicity clause, the second MalwareScan will
fail be created.

Removing the unicity clause means we need to adapt
MalwareScanRetrieveStateJob. Previously we find_by uuid,
this might apply the job to the wrong MalwareScan.

We now pass MalwareScan id to MalwareScanRetrieveStateJob
instead of UUID.