Change prefilled data banner logic for single page form
We're going to introduce a single page form with some prefilled data not
displayed. Because of this, old behavior used to displayed the banner,
without prefilled data, which is a bit confusing.

This new way has some default too: we have to manually add the new
partial on each single page view. I think the tradeoff is OK because
we'll copy some views, which embed this partial. Moreover it'll be some
generator in the future.

Ref https://github.com/etalab/data_pass/pull/500#issuecomment-2444427915