checked in radio buttons can be a lambda
I needed to be able to give a lambda to check if a radio button is checked when the value of the radio button is actually an array. In that case, rails doesn't understand if the value is in the array or not.

This should be a checkbox behavior not a radio button, but in a particular case I need to implement radio buttons that fill an array of value with a single value.