Introduce instructor draft request feature
This feature allows instructors to create draft requests to send to
applicants.

Instructors can choose a form, fill all request data within a single page
and invite applicants to claim the request.

The invite process send an email to the applicant, redirect them
to a page with a secure link, then the applicant can claim this
request, with all data filled by the instructor.

If a draft has already been claimed, or if current user does not match
the applicant, user can't claim the request.

This commit embed the whole feature (models, organizers, views) for both
instructors and users.