Harden CancelAuthorizationReopening on disjonction with snapshot
There is some cases where authorization request can be mass updated for some
reason, but not the snapshots (authorization model) (should be fixed).

For some mass update, we can remove and add keys (for example a
migration of a contact from one type to another) : on these cases,
if we try to just replace the `data` key of the authorization request
with the snapshot, we override potential some required key with a null
value : in this case, the authorization request is no longer valid, and
the organizer did not handle this edge case well.

The real life example/bug: we migrate for API Particulier the
"responsable traitement" to "contact metier", with the old behaviour we
erased old infos regarding the "contact metier" on the authorization
request if it was validated before the mass update
(because fields were not present on the snapshot), which leads to a fail
of the organizer, but did not displayed any message.

Closes https://linear.app/pole-api/issue/DAT-774