API Particulier v3: migrate data for MEN and FranceTravail
For Pole Emploi and MEN we only add new scopes to the definition, with a
migration to add all new scopes to any valid authorizations with at
least one scope related to the data provider.

Data migration has been tested locally, it's a bit complicated because
we are using a hstore for `data`, so values are string and we are
storing scopes as a ruby string array. Has been generated with LLMs