Harden archived event behaviour
The pending test is no longer valid and should be replaced with a cancel
action similar to the cancel reopening : we do not want to archive a
request with active authorizations.

This is a trade off, if there is some support in the future we will
reconsider to implement this feature.

Keep the pending test for the history and blames.