Rename Organization#raison_sociale to name
BEWARE there is a breaking change on webhooks/API. It does not impact
consumers so not a big deal

Prepare to make organization multi registry/countries, raison_sociale is french
specific. Use generic `name` instead.

raison_sociale was already a wrapper for denomination from INSEE sirene
so not a big deal