Move specific HubEE's methods from organization to API client
These fields were not used elsewhere. Because of the future multi
registry we refactor/simplify organization model.

- code_commune
- sigle_unite_legale
- libelle_commune
- code_postal

There is some small changes within HubEE factory but it is not really
relevant, we use a fix organization payload/siret, so it's a good
tradeoff and if something's break it will be easy to fix.