CancelAuthorizationReopening no longer validates authorization request
An authorization is a snapshot of a previous authorization request,
which can be dirty/invalid with actual context/validations rules.

It can happens in multiple scenarios:
1. Dirty data from migration: old authorization/requests were already
   dirty
2. Context changed: for example HubEE will validate phone number in the
   future, there is several authorizations without phone numbers.

We can reopen an authorization without errors, but the cancellation used
to restore to the latest authorization, which used to cause strange
errors. This commit fixes this behaviour.

For the context, we tried to mitigate the behaviour described in 1. with a
`dirty_from_v1` boolean, which disabled reopening (and avoid the issue
with the cancellation): we may revert this feature in the future, TBD.

In the case of a reopening without cancellation, the validations is not
an issue because we want to have clean requests (and authorizations) in
the current context.