Introduce impersonate mechanism within app
This feature allows admin to impersonate user's, with some basic
control: everytime an admin impersone someone, he has to give a legit
reason, and every actions perform on behalf of the user is tracked within
a specific table. The impersonate session has a maximum life time of 1
hour.

When an admin impersonate a user, there is a fixed banner on top to
remind that he is connected as this user.

For now, only one action is tracked (for testing purpose), next commit
will add more.