Fix infinite redirect on next stage's step
Moving the wizard's current step from session to medium lasting cookies we revealed
an issue causing infinite redirects on multi stage requests.

When starting the next stage, we tried displaying the summary which in
turn tryied to display the "wicked_finish" step which cycled back to the
summary.

This PR fixes this issue by clearing the cookie when "submitting" the
form so wicked cannot get lost.