StartNextAuthorizationRequestStage copies latest production data if exists
In the case of a full cycle approved (both sandbox S1 and production P1), and
when a user reopens in the following order:

1. A production P2
2. A sandbox S2

If user starts a new cycle from S2, data from P2 won't be within S2 next
stage (which will become P3).

In order to prevent this, when we start a new production stage we always
copies latest production data.