Introduce acceptance test for request forms steps
Will be the new way to test that forms are valid (instead of E2E tests).

We currently use E2E tests on each forms, which takes a looooonnnng time
to run. This test reproduced the step by step validation in order to
detect if a required field/step is missing.
When basic blocks are not overrided we do not need to test the steps
of our forms through E2E.

For forms/definitions with custom behaviour we do need to test views and
behaviour, these tests won't be delete.