Harden/bugfix infinite loop on dirty authorization requests
On reopening for some requests with invalid/dirty data (invalid url,
change of validation like FranceConnect mobile phone number), the
request is not submittable/reviewable, because it does not pass
validation. It is an issue for old requests, or if we change rules on
validations.

There is no rationnal reason to "block" the summary page for reopening
(which leads to an infinite loop reported multiple times by support),
especially that summary page on submit renders errors, which is good
enough to fix an issue which is a recurring one for the support.