Harden invalid reopening views within authorization request mailer
Within BaseNotifier, we use email_notification_with_reopening method to
send emails, based on request status and event. If it's a reopening, we
use the template `reopening_#{event}`, not the `#{event}` template.

Our approve custom views used to make conditions on the reopening flag within
the `approve` view, but in case of a reopening, system lookup for a
`reopening_approve`, which never exist for all of our definition.

So we create all reopening_approve custom views and rearrange approve
views.

Closes https://linear.app/pole-api/issue/DP-1316